msal: Move the application reqs from misc to msal::application
