+++ src/msal/src/lib.rs
@@ -30,3 +30,6 @@ pub mod constants;
 
 #[cfg(target_family = "unix")]
 pub mod user;
+
+#[cfg(target_family = "unix")]
+pub mod application;
